qcacld-2.0: Fill STATION_INFO_SIGNAL when returning cached value to upperlayer
When there is an get_station request from the upper layer during roaming,
the driver returns a cached rssi value but it dosent set the "sinfo->filled"
(station_info_flags) bit flag for the NL80211_STA_INFO_SIGNAL. If this flag
is not set the kernel dosen't update the cached rssi value to upper layer.

To mitigate this issue, set the "sinfo->filled" when returning the
cached value to the upper layer during roaming.

CRs-Fixed: 1084502
Change-Id: If62f7482c8153131daa2d0ea83596e5cd4e3f104
Bug: 32394855
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>